+++ modules/exercise/index.php
@@ -73,24 +73,36 @@ if ($is_editor) {
         if ($objExerciseTmp->read($exerciseId)) {
             switch ($_GET['choice']) {
                 case 'delete': // deletes an exercise
-                    $objExerciseTmp->delete();
-                    Indexer::queueAsync(Indexer::REQUEST_REMOVE, Indexer::RESOURCE_EXERCISE, $exerciseId);
-                    Session::Messages($langPurgeExerciseSuccess, 'alert-success');
-                    redirect_to_home_page('modules/exercise/index.php?course='.$course_code);
+                    if (!resource_belongs_to_progress_data(MODULE_ID_EXERCISE, $exerciseId)) {
+                        $objExerciseTmp->delete();
+                        Indexer::queueAsync(Indexer::REQUEST_REMOVE, Indexer::RESOURCE_EXERCISE, $exerciseId);
+                        Session::Messages($langPurgeExerciseSuccess, 'alert-success');
+                        redirect_to_home_page('modules/exercise/index.php?course='.$course_code);
+                    } else {
+                        Session::Messages($langResourceBelongsToCert, "alert-warning");
+                    }
                 case 'purge': // purge exercise results
-                    $objExerciseTmp->purge();
-                    Session::Messages($langPurgeExerciseResultsSuccess);
-                    redirect_to_home_page('modules/exercise/index.php?course='.$course_code);
+                    if (!resource_belongs_to_progress_data(MODULE_ID_EXERCISE, $exerciseId)) {
+                        $objExerciseTmp->purge();
+                        Session::Messages($langPurgeExerciseResultsSuccess);
+                        redirect_to_home_page('modules/exercise/index.php?course='.$course_code);
+                    } else {
+                        Session::Messages($langResourceBelongsToCert, "alert-warning");
+                    }
                 case 'enable':  // enables an exercise
                     $objExerciseTmp->enable();
                     $objExerciseTmp->save();
                     Indexer::queueAsync(Indexer::REQUEST_STORE, Indexer::RESOURCE_EXERCISE, $exerciseId);
                     redirect_to_home_page('modules/exercise/index.php?course='.$course_code);
                 case 'disable': // disables an exercise
-                    $objExerciseTmp->disable();
-                    $objExerciseTmp->save();
-                    Indexer::queueAsync(Indexer::REQUEST_STORE, Indexer::RESOURCE_EXERCISE, $exerciseId);
-                    redirect_to_home_page('modules/exercise/index.php?course='.$course_code);
+                    if (!resource_belongs_to_progress_data(MODULE_ID_EXERCISE, $exerciseId)) {
+                        $objExerciseTmp->disable();
+                        $objExerciseTmp->save();
+                        Indexer::queueAsync(Indexer::REQUEST_STORE, Indexer::RESOURCE_EXERCISE, $exerciseId);
+                        redirect_to_home_page('modules/exercise/index.php?course='.$course_code);
+                    } else {
+                        Session::Messages($langResourceBelongsToCert, "alert-warning");
+                    }
                 case 'public':  // make exercise public
                     $objExerciseTmp->makepublic();
                     $objExerciseTmp->save();
